Implemented the Simplified User Interface with Re-Run Simulation Feature
Site Name
Second Input entry should be Unit Name
3rd Entry is Owner's Name
Output file name is a combination of SiteName and Unit Name
Now output file is having owner's Name.  Check what all details should
be ther e in the output file.
Today's Date should be Removed and the application should consider the
Todays' date by default.
On Parennial selection the start and end dates are greyed.
The Parennial block should be shifted above the start and end date
What should be the default name of the Unit, NO Default Name, Just Empty
Month in the begin and end dates , I have given the Month and Day as the
default Option and Left the User to select this. Validation is give to
for the user to select the date.
Ask for Clarification on the Output Style
On Re-Run Button it should go back to the first frame with the same
memory
On New Project It should run the new Project with clear memory.
There are 3 buttons now, New Project, Summary, Re-run